Ignored lots of compiler warnings.
Mostly these are deprecations from the Spigot team which I believe
shouldn't be deprecated.  For example, players refer to each other by
name, not UUID - so there will always be a need for player lookup by
name.  Also the block IDs are a well-documented standard that everyone
understands, even if they're not very human-friendly.  Plugins use those
IDs and data values to specify block types for example in config files.
As for the rest of the ignores, I either decided the warnings are just
noise based on the situation, or that I'm comfortable with the risks.
Possibly for the first time in 5 years of dev work on this plugin, I
just compiled without any warnings.  :)