Handle EntityCombustByEntity events.
A stab at refactoring to handle both damage and combustion together without code
duplication.

It's a mess so far.  Please forgive me, I'm not a Java programmer!

If I get it to work, I'll try to tidy it up a bit.

Choking so far on errors that seem unrelated to what I've done - several
instances of:

 incompatible types: org.bukkit.entity.Player cannot be converted to
java.lang.String

... but I can't see why.