Handle EntityCombustByEntityEvent properly.
This makes it actually prevent normal damage and combustion, for me.

I want to test it further on my test server, but going to raise a PR to share my
work so far.

This may not be the best way to do this - I re-iterate that I'm not a Java dev -
and it needs tidying, indentation seems to be all over the place (but then I
indent with 4 spaces per level, I think the original code may have used tabs; if
this approach is acceptable to the GP devs, I'm happy to work on tidying it up a
bit, but would like feedback first.