Disable delayed logout messages by default.
This feature has value in anti-spam, but too many new GP installers are
confused by it.  Disabling by default to save support cost - also I
figure at least some are installing the plugin, seeing this as a "bug"
and uninstalling without bothering to get educated, and that's bad for
their players.